// src/app/Models/property.model.ts
export class Property {
	name: string; //Ok
	rooms: number; //Ok
	restroom: number; 
	lotarea: number; //Ok
	builtarea: number; //OK
	pools: boolean;
	parking: boolean;
	guestparking: boolean;
	price: number; 
	gym: boolean;
	location: number; //Ok
    floors : number; //Ok
	client:string; //Ok
    
}
/*
	Id : number;
	Name: string;
	Rooms: number;
	Bathroom: number;
	Type: number;
	LotArea: number;
	BuiltArea: number;
	Pools: boolean;
	Parkings: boolean;
	Price: number;
	Gym: boolean;
	Location: number;
	Description: string;
    Floors : number; 
	Client:string;
	
*/